updgrade to golioth sdk v0.3.0
After upgrading, for some reason the MTB build process
did not include the mtb-hal-cat1 headers or source files,
even though that library is included in the libs dir.
To work around this, I manually included the include dirs
and sources to the app Makefiles.

Also, updated .cyignore for v0.3.0 of the Golioth SDK
to exclude source files not intended for MTB.

Signed-off-by: Nick Miller <nick@golioth.io>