deps: use wifi-core-freertos-lwip-mbedtls.mtb
Recent versions of ModusToolbox (3+) use a newer bundled .mtb for
WiFi driver, FreeRTOS, LwIP, and MbedTLS.

The prior wifi-connection-manager.mtb no longer works in MTB 3.0.
When the WiFi driver is initialized, the system can become unstable,
and the serial output is garbled. The root cause is unknown,
but the problem does not exist in the new .mtb.

Signed-off-by: Nick Miller <nick@golioth.io>